requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
bson==0.5.10
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
geographiclib==2.0
geopy==2.3.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
redis==4.4.2
SQLAlchemy==2.0.0
typing_extensions==4.4.0
